backend/accounts/views.py
from rest_framework import status
from rest_framework.response import Response
from rest_framework.views import APIView
from rest_framework_simplejwt.views import TokenObtainPairView
from rest_framework_simplejwt.tokens import RefreshToken
from .serializers import UserSerializer, CustomTokenObtainPairSerializer
from rest_framework.permissions import AllowAny
from rest_framework.exceptions import AuthenticationFailed
from django.contrib.auth import get_user_model
from django.utils.encoding import force_bytes
from django.utils.http import urlsafe_base64_encode
from django.contrib.auth.tokens import default_token_generator
from django.conf import settings
from django.core.mail import send_mail

class RegisterView(APIView):
    permission_classes = [AllowAny]

    def post(self, request):
        serializer = UserSerializer(data=request.data)
        if serializer.is_valid():
            user = serializer.save()
            refresh = RefreshToken.for_user(user)
            access_token = str(refresh.access_token)

            user_data = {
                "mobile_number": user.mobile_number,
                "email": user.email,
                "full_name": user.get_full_name(),
                "role": user.role,
                "district": user.district,
                "state": user.state,
                "pincode": user.pincode,
                "batch_name": user.batch_name,
                "subjects": user.subjects,
            }

            return Response(
                {
                    "tokens": {"refresh": str(refresh), "access": access_token},
                    "user": user_data,
                },
                status=status.HTTP_201_CREATED,
            )
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)

class CustomTokenObtainPairView(TokenObtainPairView):
    serializer_class = CustomTokenObtainPairSerializer

    def post(self, request, *args, **kwargs):
        try:
            response = super().post(request, *args, **kwargs)
            if response.status_code == 200:
                response.data['message'] = "Login successful! Welcome back."
            return response
        except AuthenticationFailed as exc:
            message = str(exc)
            if "No active account found" in message:
                return Response({"error": "Invalid mobile number or password"}, status=status.HTTP_401_UNAUTHORIZED)
            if "inactive" in message.lower():
                return Response({"error": "This account has been deactivated"}, status=status.HTTP_403_FORBIDDEN)
            return Response({"error": "Invalid mobile number or password"}, status=status.HTTP_401_UNAUTHORIZED)
        except Exception:
            return Response({"error": "Login failed. Please try again."}, status=status.HTTP_400_BAD_REQUEST)

class PasswordResetRequestView(APIView):
    permission_classes = [AllowAny]

    def post(self, request):
        email = request.data.get("email")
        if not email:
            return Response({"email": ["This field is required."]}, status=status.HTTP_400_BAD_REQUEST)

        User = get_user_model()
        try:
            user = User.objects.get(email=email)
        except User.DoesNotExist:
            # For security, respond success even if user not found
            return Response({"detail": "If an account with that email exists, a reset link has been sent."})

        uidb64 = urlsafe_base64_encode(force_bytes(user.pk))
        token = default_token_generator.make_token(user)

        frontend_base = getattr(settings, "FRONTEND_BASE_URL", "http://localhost:5173")
        reset_link = f"{frontend_base}/reset-password/{uidb64}/{token}"

        subject = "Password Reset Request"
        message = f"Use the following link to reset your password: {reset_link}"

        try:
            send_mail(subject, message, getattr(settings, "DEFAULT_FROM_EMAIL", "no-reply@example.com"), [email])
        except Exception:
            # Even if email fails in dev, return success to avoid enumeration
            pass

        return Response({"detail": "If an account with that email exists, a reset link has been sent."})
