backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-browser-reload==1.21.0
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
Pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
pymongo[srv]==4.9.1 